requirements.txt
mysql-connector-python==8.1.0
pandas==2.1.0
pyodbc==4.0.39
